backend/requirements.txt
fastapi
uvicorn
phidata
groq
python-dotenv
requests
sqlalchemy
apscheduler
duckduckgo-search
jinja2
xhtml2pdf
pocketbase
